rename AzureB2C to AzureAD and scaffold new files with more airbnb data

+++ inside-airbnb/Models/GeoJson/Property.cs
@@ -2,11 +2,11 @@
 {
     public class Property
     {
-        public Property(int listingId)
+        public Property(long listingId)
         {
             ListingId = listingId;
         }
 
-        public int ListingId { get; set; }
+        public long ListingId { get; set; }
     }
 }
