+++ inside-airbnb/Services/IListingService.cs
@@ -5,9 +5,9 @@ namespace inside_airbnb.Services
     public interface IListingService
     {
         Task<IEnumerable<Listing>> GetListings();
-        Task<Listing> GetListingByID(int listingId);
+        Task<Listing> GetListingByID(long listingId);
         Task InsertListing(Listing listing);
-        Task DeleteListing(int listingId);
+        Task DeleteListing(long listingId);
         Task UpdateListing(Listing listing);
         Task Save();
     }
